Organize custom dimension functions in a separate file
